move strategy to correct location, add tests for getitem
