+++ tests/custom_strategies/__init__.py
@@ -1,7 +1,9 @@
 """ Custom hypothesis search strategies """
 import math
 from functools import lru_cache, partial
+from itertools import groupby
 from numbers import Integral
+from operator import itemgetter
 from typing import Any, Iterable, List, Optional, Tuple, Union
 
 import hypothesis.extra.numpy as hnp
@@ -358,3 +360,148 @@ def valid_shapes(
         if len(shape) == 1 and draw(st.booleans())
         else tuple(int(i) for i in shape)
     )
+
+
+@st.composite
+def arbitrary_indices(draw, shape):
+    """
+    Hypothesis search strategy: Generate a valid index
+    for an array of a given shape. The index can contain
+    any type of valid object used for indexing, including
+    integers, slices, Ellipsis's, newaxis's, boolean arrays,
+    and integer arrays.
+
+    Parameters
+    ----------
+    shape : Tuple[int]
+        The shape of the array to be indexed into
+
+    Notes
+    -----
+    `draw` is a parameter reserved by hypothesis, and should not be specified
+    by the user.
+
+    Returns
+    -------
+    hypothesis.searchstrategy.SearchStrategy[Tuple[Union[int, slice, Ellipsis, NoneType, numpy.ndarray], ...]]
+    """
+    def group_continuous_integers(ls):
+        """
+        Given a list of integers, find and group continuous sequences
+
+        Parameters
+        ----------
+        ls: List[int]
+
+        Returns
+        -------
+        List[Tuple[int]]
+
+        Examples
+        --------
+        >>> group_continuous_integers([1, 3, 4, 5, 7, 8])
+        [(1,), (3, 4, 5), (7, 8)]
+        """
+        return [
+            tuple(map(itemgetter(1), g))
+            for k, g in groupby(enumerate(ls), lambda x: x[0] - x[1])
+        ]
+
+    shape_inds = list(range(len(shape)))
+    index = []  # stores tuples of (dim, index)
+
+    # add integers, slices
+    basic_inds = sorted(draw(st.lists(st.sampled_from(shape_inds), unique=True)))
+
+    if len(basic_inds) > 0:
+        basic_dims = tuple(shape[i] for i in basic_inds)
+
+        # only draw ints and slices
+        # will handle possible ellipsis and newaxis objects later
+        # as these can make boolean indices difficult to handle later
+        basics = draw(hnp.basic_indices(shape=basic_dims, allow_ellipsis=False))
+        if not isinstance(basics, tuple):
+            basics = (basics,)
+
+        # will not necessarily index all dims from basic_inds as
+        # `basic_indices` can return indices with omitted trailing slices
+        index += [tup for tup in zip(basic_inds, basics)]
+
+        for i in basic_inds[: len(basics)]:
+            shape_inds.pop(shape_inds.index(i))
+
+    if len(shape_inds) > 0:
+        # add integer arrays to index
+        int_arr_inds = sorted(draw(st.lists(st.sampled_from(shape_inds), unique=True)))
+
+        if len(int_arr_inds) > 0:
+            int_arr_dims = tuple(shape[i] for i in int_arr_inds)
+            int_arrs = draw(hnp.integer_array_indices(shape=int_arr_dims))
+            index += [tup for tup in zip(int_arr_inds, int_arrs)]
+
+            for i in int_arr_inds:
+                shape_inds.pop(shape_inds.index(i))
+
+        if len(shape_inds) > 0:
+            # add boolean arrays to index
+            bool_inds = sorted(draw(st.lists(st.sampled_from(shape_inds), unique=True)))
+
+            if len(bool_inds) > 0:
+                grouped_bool_inds = group_continuous_integers(bool_inds)
+                bool_dims = [tuple(shape[i] for i in ind) for ind in grouped_bool_inds]
+
+                # if multiple boolean array indices, the number of trues must be the same
+                # such that the output of ind.nonzero() for each index are broadcast compatible
+                # this must also be the same as the trailing index of each integer array, if any used
+                if len(int_arr_inds):
+                    max_trues = max(i.shape[-1] for i in int_arrs)
+                else:
+                    max_trues = st.integers(
+                        min_value=0, max_value=min(bool_dims, key=lambda x: np.prod(x))
+                    )
+
+                index += [
+                    (
+                        i[0],
+                        draw(
+                            hnp.arrays(shape=sh, dtype=bool).filter(
+                                lambda x: x.sum() in (1, max_trues)
+                            )
+                        ),
+                    )
+                    for i, sh in zip(grouped_bool_inds, bool_dims)
+                ]
+
+                for i in bool_inds:
+                    shape_inds.pop(shape_inds.index(i))
+
+    grouped_shape_inds = group_continuous_integers(sorted(shape_inds))
+    if len(grouped_shape_inds) == 1:
+        # unused indices form a continuous stretch of dimensions
+        # so can replace with an ellipsis
+
+        # to test ellipsis vs omitted slices
+        # randomly select if the unused indices are trailing
+        if max(shape_inds) + 1 == len(shape):
+            if draw(st.booleans()):
+                index += [(min(shape_inds), Ellipsis)]
+        else:
+            index += [(min(shape_inds), Ellipsis)]
+
+    else:
+        # so that current chosen index's work,
+        # fill in remaining any gaps with empty slices
+        index += [(i, slice(None)) for i in shape_inds]
+
+    index = sorted(index, key=lambda x: x[0])
+    out_ind = tuple(i[1] for i in index)
+
+    # can now randomly add in newaxis objects
+    newaxis_pos = sorted(
+        draw(st.lists(st.integers(min_value=0, max_value=len(index)), unique=True)),
+        reverse=True,
+    )
+    for i in newaxis_pos:
+        index.insert(i, np.newaxis)
+
+    return out_ind
